requirements.txt
ibapi==9.81.1.post1
python-dotenv==1.0.1
pydantic==2.10.6
